# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    VirtualFileEntry:
      type: object
      required: [path, name, isDirectory]
      properties:
        path:
          type: string
        name:
          type: string
        isDirectory:
          type: boolean
        size:
          type: integer
        children:
          type: array
          items:
            $ref: "#/components/schemas/VirtualFileEntry"
...
    KanbanColumnAutomation:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        providerId:
          type: string
        role:
          type: string
        specialistId:
          type: string
        specialistName:
          type: string
        transitionType:
          type: string
          enum: [entry, exit, both]
        requiredArtifacts:
          type: array
          items:
            type: string
            enum: [screenshot, test_results, code_diff]
        autoAdvanceOnSuccess:
          type: boolean
    KanbanColumn:
      type: object
      required: [id, name, position, stage]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        position:
          type: integer
        stage:
          type: string
        automation:
          $ref: "#/components/schemas/KanbanColumnAutomation"
    KanbanBoard:
      type: object
      required: [id, workspaceId, name, isDefault, columns, createdAt, updatedAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        isDefault:
          type: boolean
        columns:
          type: array
          items:
            $ref: "#/components/schemas/KanbanColumn"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/kanban/boards:
    get:
      operationId: listKanbanBoards
      summary: List Kanban boards for a workspace
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: string
            default: "default"
      responses:
        "200":
          description: Board list
          content:
            application/json:
              schema:
                type: object
                properties:
                  boards:
                    type: array
                    items:
                      $ref: "#/components/schemas/KanbanBoard"
    post:
      operationId: createKanbanBoard
      summary: Create a Kanban board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId, name]
              properties:
                workspaceId:
                  type: string
                name:
                  type: string
                isDefault:
                  type: boolean
      responses:
        "201":
          description: Board created
          content:
            application/json:
              schema:
                type: object
                properties:
                  board:
                    $ref: "#/components/schemas/KanbanBoard"
...
  /api/clone/local:
    post:
      operationId: loadLocalRepo
      summary: Load an existing local git repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
      responses:
        "200":
          description: Loaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  name:
                    type: string
                  path:
                    type: string
                  branch:
                    type: string
                  branches:
                    type: array
                    items:
                      type: string
...
  /api/github/import:
    post:
      operationId: importGitHubRepo
      summary: Import a GitHub repo as a virtual workspace (zipball download)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                owner:
                  type: string
                repo:
                  type: string
                ref:
                  type: string
                  description: Branch, tag, or SHA. Defaults to HEAD.
                url:
                  type: string
                  description: Shorthand — GitHub URL (parsed into owner/repo)
      responses:
        "200":
          description: Import successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  owner:
                    type: string
                  repo:
                    type: string
                  ref:
                    type: string
                  fileCount:
                    type: integer
                  extractedPath:
                    type: string
                  importedAt:
                    type: string
                    format: date-time
        "404":
          description: Repository not found
        "403":
          description: Rate limited or forbidden
        "413":
          description: Repository too large
...
  /api/github/tree:
    get:
      operationId: getGitHubTree
      summary: Get file tree for an imported GitHub repo
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
        - name: repo
          in: query
          required: true
          schema:
            type: string
        - name: ref
          in: query
          schema:
            type: string
            default: HEAD
      responses:
        "200":
          description: File tree
          content:
            application/json:
              schema:
                type: object
                properties:
                  tree:
                    type: array
                    items:
                      $ref: "#/components/schemas/VirtualFileEntry"
                  fileCount:
                    type: integer
        "404":
          description: Workspace not imported
  /api/github/file:
    get:
      operationId: getGitHubFile
      summary: Read a file from an imported GitHub repo
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
        - name: repo
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: ref
          in: query
          schema:
            type: string
            default: HEAD
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                  path:
                    type: string
        "404":
          description: File or workspace not found
  /api/github/search:
    get:
      operationId: searchGitHubFiles
      summary: Search files in an imported GitHub repo
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
        - name: repo
          in: query
          required: true
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: ref
          in: query
          schema:
            type: string
            default: HEAD
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        name:
                          type: string
                        score:
                          type: number
                  total:
                    type: integer
                  query:
                    type: string

  # ── Workspace sub-resources ──
...